Rename pino to structuredLogger
