+++ packages/api/src/router/generations.ts
@@ -13,7 +13,7 @@ import {
 } from "@oakai/core/src/types";
 import { sendQuizFeedbackEmail } from "@oakai/core/src/utils/sendQuizFeedbackEmail";
 import { requestGenerationWorker } from "@oakai/core/src/workers/generations/requestGeneration";
-import { legacyLogger as logger } from "@oakai/logger";
+import { structuredLogger as logger } from "@oakai/logger";
 import { aiLogger } from "@oakai/logger";
 import { TRPCError } from "@trpc/server";
 import { Redis } from "@upstash/redis";
