+++ packages/core/src/models/generations.ts
@@ -5,7 +5,7 @@ import {
   Prisma,
   PrismaClientWithAccelerate,
 } from "@oakai/db";
-import { legacyLogger, LegacyLogger } from "@oakai/logger";
+import { structuredLogger, StructuredLogger } from "@oakai/logger";
 import { Logger as InngestLogger } from "inngest/middleware/logger";
 import { omit } from "remeda";
 import { Md5 } from "ts-md5";
@@ -22,7 +22,9 @@ export class Generations {
     private readonly prisma: PrismaClientWithAccelerate,
     // inngest's logger doesn't allow child logger creation, so make
     // sure we accept instances of that too
-    private readonly logger: LegacyLogger | InngestLogger = legacyLogger,
+    private readonly logger:
+      | StructuredLogger
+      | InngestLogger = structuredLogger,
   ) {}
 
   async byId(generationId: string): Promise<Generation | null> {
